feat: サポカの限界突破レベルの表示/変更を行う箇所を新たなコンポーネントに分割
